final review of Zhen's code
