+++ README.md
@@ -1,5 +1,5 @@
 # amr-qa
-This repo maps natural language questions to SPARQL queries via AMR representations
+This repo contains procedures that map natural language questions to SPARQL queries via AMR representations.
 
 ## Configuring the environment
 ```
@@ -8,6 +8,12 @@ python3 -m venv .venv
 source .venv/bin/activate
 pip install -r requirements.txt
 ```
+
+## Downloading data
+
+## Querying the SPARQL API
+
+## Evaluate the system
 
 ## Run tests
 ```
